requirements.txt
# Core dependencies
watchdog==3.0.0
PyYAML==6.0.1
requests==2.31.0
python-dotenv==1.0.0

# PDF processing
PyPDF2==3.0.1
pdfplumber==0.10.3
reportlab==4.0.7
weasyprint==60.2

# OCR capabilities
pytesseract==0.3.10
Pillow==10.1.0
pdf2image==1.16.3

# Markdown processing
markdown==3.5.1
pypandoc==1.11

# Email functionality (smtplib is built-in)

# File handling (pathlib and shutil are built-in)

# Logging and utilities
loguru==0.7.2
click==8.1.7

# Phase 1: Security & Reliability
cryptography==41.0.7
tenacity==8.2.3
pydantic==2.5.0
python-magic==0.4.27

# Phase 2: Performance & Architecture
aiofiles==23.2.1
aiohttp==3.9.1
aiosmtplib==3.0.1
sqlalchemy==2.0.23
alembic==1.13.1
prometheus-client==0.19.0

# Phase 3: Quality & Operations
black==23.12.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
pytest-cov==4.1.0
pytest-benchmark==4.0.0
pre-commit==3.6.0

# Phase 4: Advanced Features
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation==0.42b0
opentelemetry-instrumentation-aiohttp-client==0.42b0
opentelemetry-instrumentation-sqlalchemy==0.42b0
opentelemetry-instrumentation-requests==0.42b0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-exporter-otlp==1.21.0
redis==5.0.1
slowapi==0.1.9
